requirements-core.txt
# --- Core Project Libraries ---
# apache-airflow==2.7.3
flask>=2.2.1,<2.3
yfinance==0.2.63
fredapi==0.5.2
pandas==2.3.0
numpy==2.2.6
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.7.0
# prefect==3.4.6
jupyter==1.1.1

# --- Logging ---
loguru==0.7.2

# --- Telemetry & Monitoring ---
prometheus-client==0.22.1
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1

# --- Visualization ---
plotly==5.22.0
altair==5.3.0

# --- Linting & Formatting ---
black==24.4.2
flake8==7.0.0
isort==5.13.2

# --- Type Checking ---
mypy==1.10.0

# --- Environment Management ---
python-dotenv==1.0.1

# --- Testing ---
pytest==8.2.2
pytest-cov==5.0.0

# --- Development Utilities ---
ipython==8.25.0
ipython==8.25.0
